BAEL-169: Refactored controller into their own classes. Renamed modules.
