+++ spring-cloud-hystrix/feign-rest-consumer/src/main/java/com/baeldung/spring/cloud/hystrix/rest/consumer/RestConsumerFeignApplication.java
@@ -1,0 +1,17 @@
+package com.baeldung.spring.cloud.hystrix.rest.consumer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.circuitbreaker.EnableCircuitBreaker;
+import org.springframework.cloud.netflix.feign.EnableFeignClients;
+import org.springframework.cloud.netflix.hystrix.dashboard.EnableHystrixDashboard;
+
+@SpringBootApplication
+@EnableCircuitBreaker
+@EnableHystrixDashboard
+@EnableFeignClients
+public class RestConsumerFeignApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(RestConsumerFeignApplication.class, args);
+    }
+}
