+++ spring-cloud-hystrix/feign-rest-consumer/src/main/resources/application.properties
