+++ spring-cloud-hystrix/feign-rest-consumer/src/main/resources/templates/greeting-view.html
