+++ spring-cloud-hystrix/pom.xml
@@ -7,14 +7,13 @@
     <artifactId>spring-cloud-hystrix</artifactId>
     <version>1.0.0-SNAPSHOT</version>
     <modules>
-        <module>spring-cloud-hystrix-rest-producer</module>
-        <module>spring-cloud-hystrix-rest-consumer</module>
-        <module>spring-cloud-hystrix-feign-rest-consumer</module>
+        <module>rest-producer</module>
+        <module>rest-consumer</module>
+        <module>feign-rest-consumer</module>
     </modules>
     <packaging>pom</packaging>
 
-    <name>Spring Cloud Hystrix</name>
-    <description>Spring Cloud Hystrix Demo</description>
+    <name>spring-cloud-hystrix</name>
 
     <parent>
         <groupId>com.baeldung</groupId>
