+++ spring-cloud-hystrix/rest-consumer/src/main/java/com/baeldung/spring/cloud/hystrix/rest/consumer/GreetingController.java
@@ -1,0 +1,19 @@
+package com.baeldung.spring.cloud.hystrix.rest.consumer;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+public class GreetingController {
+    @Autowired
+    private GreetingService greetingService;
+
+    @RequestMapping("/get-greeting/{username}")
+    public String getGreeting(Model model, @PathVariable("username") String username) {
+        model.addAttribute("greeting", greetingService.getGreeting(username));
+        return "greeting-view";
+    }
+}
