+++ spring-cloud-hystrix/rest-consumer/src/main/java/com/baeldung/spring/cloud/hystrix/rest/consumer/RestConsumerApplication.java
@@ -1,0 +1,15 @@
+package com.baeldung.spring.cloud.hystrix.rest.consumer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.circuitbreaker.EnableCircuitBreaker;
+import org.springframework.cloud.netflix.hystrix.dashboard.EnableHystrixDashboard;
+
+@SpringBootApplication
+@EnableCircuitBreaker
+@EnableHystrixDashboard
+public class RestConsumerApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(RestConsumerApplication.class, args);
+    }
+}
