+++ spring-cloud-hystrix/rest-consumer/src/main/resources/application.properties
