+++ spring-cloud-hystrix/rest-consumer/src/main/resources/templates/greeting-view.html
