+++ spring-cloud-hystrix/rest-producer/src/main/resources/application.properties
