--- spring-cloud-hystrix/spring-cloud-hystrix-feign-rest-consumer/src/main/java/com/baeldung/spring/cloud/hystrix/rest/consumer/RestConsumerFeignApplication.java
@@ -1,32 +1,0 @@
-package com.baeldung.spring.cloud.hystrix.rest.consumer;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.circuitbreaker.EnableCircuitBreaker;
-import org.springframework.cloud.netflix.feign.EnableFeignClients;
-import org.springframework.cloud.netflix.hystrix.dashboard.EnableHystrixDashboard;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@SpringBootApplication
-@EnableCircuitBreaker
-@EnableHystrixDashboard
-@EnableFeignClients
-@Controller
-public class RestConsumerFeignApplication {
-    @Autowired
-    private GreetingClient greetingClient;
-
-    public static void main(String[] args) {
-        SpringApplication.run(RestConsumerFeignApplication.class, args);
-    }
-
-    @RequestMapping("/get-greeting/{username}")
-    public String getGreeting(Model model, @PathVariable("username") String username) {
-        model.addAttribute("greeting", greetingClient.greeting(username));
-        return "greeting-view";
-    }
-}
