--- spring-cloud-hystrix/spring-cloud-hystrix-rest-consumer/src/main/java/com/baeldung/spring/cloud/hystrix/rest/consumer/RestConsumerApplication.java
@@ -1,31 +1,0 @@
-package com.baeldung.spring.cloud.hystrix.rest.consumer;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.circuitbreaker.EnableCircuitBreaker;
-import org.springframework.cloud.netflix.hystrix.EnableHystrix;
-import org.springframework.cloud.netflix.hystrix.dashboard.EnableHystrixDashboard;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@SpringBootApplication
-@EnableCircuitBreaker
-@EnableHystrixDashboard
-@Controller
-public class RestConsumerApplication {
-    @Autowired
-    private GreetingService greetingService;
-
-    public static void main(String[] args) {
-        SpringApplication.run(RestConsumerApplication.class, args);
-    }
-
-    @RequestMapping("/get-greeting/{username}")
-    public String getGreeting(Model model, @PathVariable("username") String username) {
-        model.addAttribute("greeting", greetingService.getGreeting(username));
-        return "greeting-view";
-    }
-}
