--- spring-cloud-hystrix/spring-cloud-hystrix-rest-producer/src/main/java/com/baeldung/spring/cloud/hystrix/rest/producer/RestProducerApplication.java
@@ -1,20 +1,0 @@
-package com.baeldung.spring.cloud.hystrix.rest.producer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-@SpringBootApplication
-@RestController
-public class RestProducerApplication implements GreetingController {
-    public static void main(String[] args) {
-        SpringApplication.run(RestProducerApplication.class, args);
-    }
-
-    @Override
-    public String greeting(@PathVariable("username") String username) {
-        return String.format("Hello %s!\n", username);
-    }
-}
